- separate help modal function to new file - implement help modal for selector
